Continued progress on Burtsev model
Added features and got all the logic working like in the original paper.
Increased speed slightly. Will need to increase speed if we want to be
able to run the simulation for 7 digit number of iterations.

Additionally from the IPD model got the html to read json from a file so
no more hardcoding in the graphs, this will allow us to easily add in
historical snapshots once the new agent structure is completed.